Address PR #727 review feedback
Fix 3 bugs and 2 minor issues from code review:

1. Use actual dt parameter instead of hardcoded 0.016 for gallery
   smooth scroll interpolation — fixes inaccurate animation on
   non-60Hz displays
2. Right-align unlock count in gallery header per spec — was
   incorrectly left-aligned at CONTENT_PADDING, now positioned at
   PANEL_W - CONTENT_PADDING
3. Fix progress percentage text overflow past card boundary — reduce
   bar width by 40px and right-align the percentage label at the
   card edge (x + CARD_W - 8) instead of barX + barW + 4
4. Extract shared formatRelativeDate utility from duplicated logic
   in HUD.formatSaveDate and AchievementGallery.formatDate
5. Add consistent break after openAchievementGallery() in pause
   handler to match menu handler pattern

Made-with: Cursor
